+++ src/main/javascript/webpack/BuildUtils.js
@@ -14,6 +14,21 @@ function artifactName(projectName, baseName) {
   return name;
 }
 
+function extractVendors (packageJson)
+{
+  "use strict";
+
+  let vendorPackages = [];
+  if (packageJson.dependencies) {
+    vendorPackages = Object.keys(packageJson.dependencies).concat(vendorPackages);
+  }
+  if (packageJson.devDependencies) {
+    vendorPackages = Object.keys(packageJson.devDependencies).concat(vendorPackages);
+  }
+
+  return vendorPackages;
+}
+
 module.exports = {
   artifactName: function (projectDir) {
     "use strict";
@@ -23,5 +38,11 @@ module.exports = {
     return function (baseName) {
       return artifactName(projectName, baseName);
     }
+  },
+  autoVendorPackages: function (projectDir) {
+    "use strict";
+    const packageJsonPath = path.resolve(projectDir, 'package.json');
+    const packageJson = require(packageJsonPath);
+    return extractVendors(packageJson);
   }
 };